Refactored login form to use components
